// packages/cdk/src/web/function-app-types.d.ts
export interface RedisStreamTriggerConfig {
    readonly type: 'redisStream';
    /**
     * Redis connection string app setting name.
     */
    readonly connection: string;
    /**
     * Stream key to monitor.
     */
    readonly key: string;
    /**
     * Consumer group name.
     */
    readonly consumerGroup?: string;
    /**
     * Polling interval in milliseconds.
     */
    readonly pollingIntervalInMs?: number;
    /**
     * Maximum batch size.
     */
    readonly maxBatchSize?: number;
}